pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "test-ocr-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "anthropic>=0.69.0",
    "azure-ai-documentintelligence>=1.0.2",
    "beautifulsoup4>=4.12.0",
    "chardet>=5.2.0",
    "google-genai>=1.29.0",
    "ipykernel>=6.30.0",
    "japanize-matplotlib>=1.1.3",
    "jupyterlab>=4.4.5",
    "langchain-anthropic>=0.3.21",
    "langchain-openai>=0.3.35",
    "Levenshtein>=0.26.1",
    "llama-cloud-services>=0.6.57",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "paddleocr[all]>=3.3.1",
    "paddlepaddle>=3.2.1",
    "page-dewarp>=0.2.3",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pydantic-settings>=2.0.0",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "python-levenshtein>=0.27.3",
    "qwen-vl-utils>=0.0.14",
    "requests>=2.32.4",
    "scipy>=1.16.3",
    "transformers>=4.56.2",
    "yomitoku>=0.9.5",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
]

[tool.uv]
package = true

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git" }

[project.scripts]
layout = "src.run_models:main_layout"
ocr = "src.run_models:main_ocr"
preprocess = "src.run_preprocessing:main"
postprocess = "src.run_postprocess:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
pythonpath = ["."]
